requirements.txt
flask==3.0.0
requests==2.31.0
markdown==3.4.4
python-dotenv==1.0.0